Move internal/utility to pkg/util as suggested by @chronark.
